refactor(analysis): rename modules/ to analysis/
