+++ Data/.gitignore
@@ -1,2 +1,3 @@
 **/*_check_output.txt
 **/.~lock.*
+output/**
