+++ backend/.gitignore
@@ -2,4 +2,4 @@ output/
 .venv/
 
 # generated python modules from fortran code with f2py
-ccfatigue/fortran/
+ccfatigue/analysis/fortran/*.so
