+++ backend/ccfatigue/analysis/sn_curve_linlog.py
@@ -18,7 +18,7 @@ from typing import Optional
 import pandas as pd
 from pandas._typing import FilePath, ReadBuffer, WriteBuffer
 
-import ccfatigue.modules.sn_curve_utils as sn_curve_utils
+import ccfatigue.analysis.sn_curve_utils as sn_curve_utils
 
 # Prepare all cycles to failure output (SNC)
 # Refs:
