--- backend/ccfatigue/modules/astm/README.md
@@ -1,35 +1,0 @@
-# ASTM
-
-## astm.csv 
-{Description of this dataset to be filled by Tassos!}
-
-This datasets contains values of t_a (??) in unit (??)
-
-Columns (	50	75	90	95	97.5	99	99.5) correspond to the confidence interval (??)
-
-Rows (indexed 3 to inf) correspond to (??)
-
-Source: STP313_Fittin_straight_lines.pdf, p.6.
-
-Used by methods: S-N-Curve-LogLog.py
-
-## astm{reliability_level}.csv for reliability level = 95% and 99%
-{Description of this dataset to be filled by Tassos!}
-
-This datasets contains values of (??)
-
-Columns mean (??)
-
-Rows mean (??)
-
-Source: STP313_Fittin_straight_lines.pdf, pp. 20-23, DOI: 10.1520/STP313-EB
-
-Used by methods: S-N-Curve-LogLog.py
-
-
-## ASTM STP 734 - Sendeckyj
-Fitting Models to Composite Materials Fatigue Data
-
-Source: STP734_1981.pdf, pp. 245-260, DOI: 10.1520/STP29314S
-
-Used by methods: S-N-Curve-Sendeckyj
