--- backend/ccfatigue/modules/external_data_sources/astm.csv
@@ -1,33 +1,0 @@
-﻿n,50,75,90,95,97.5,99,99.5
-3,1.00000,2.4142,6.3138,12.706,25.452,63.657,127.32
-4,0.81650,1.6036,2.9200,4.3027,6.2053,9.9248,14.089
-5,0.76489,1.4226,2.3534,3.1825,4.1765,5.8409,7.4533
-6,0.74070,1.3444,2.1318,2.7764,3.4954,4.6041,5.5976
-7,0.72669,1.3009,2.0150,2.5706,3.1634,4.0321,4.7733
-8,0.71756,1.2733,1.9432,2.4469,2.9687,3.7074,4.3168
-9,0.71114,1.2543,1.8946,2.3646,2.8412,3.4995,4.0293
-10,0.70639,1.2403,1.8595,2.3060,2.7515,3.3554,3.8325
-11,0.70272,1.2297,1.8331,2.2622,2.685,3.2498,3.6897
-12,0.69981,1.2213,1.8125,2.2281,2.6338,3.1693,3.5814
-13,0.69745,1.2145,1.7959,2.2010,2.5931,3.1058,3.4966
-14,0.69548,1.2089,1.7823,2.1788,2.56,3.0545,3.4284
-15,0.69384,1.2041,1.7709,2.1604,2.5326,3.0123,3.3725
-16,0.69242,1.2001,1.7613,2.1448,2.5096,2.9768,3.3257
-17,0.69120,1.1967,1.7530,2.1315,2.4899,2.9467,3.286
-18,0.69013,1.1937,1.7459,2.1199,2.4729,2.9208,3.252
-19,0.68919,1.1910,1.7396,2.1098,2.4581,2.8982,3.2225
-20,0.68837,1.1887,1.7341,2.1009,2.445,2.8784,3.1966
-21,0.68763,1.1866,1.7291,2.0930,2.4334,2.8609,3.1737
-22,0.68696,1.1848,1.7247,2.0860,2.4231,2.8453,3.1534
-23,0.68635,1.1831,1.7207,2.0796,2.4138,2.8314,3.1352
-24,0.68580,1.1816,1.7171,2.0739,2.4055,2.8188,3.1188
-25,0.68531,1.1802,1.7139,2.0687,2.3979,2.8073,3.104
-26,0.68485,1.1789,1.7109,2.0639,2.391,2.7969,3.0905
-27,0.68443,1.1777,1.7081,2.0595,2.3846,2.7874,3.0782
-28,0.68405,1.1766,1.7056,2.0555,2.3788,2.7787,3.0669
-29,0.68370,1.1757,1.7033,2.0518,2.3734,2.7707,3.0565
-30,0.68335,1.1748,1.7011,2.0484,2.3685,2.7633,3.0469
-40,0.68099,1.1682,1.6860,2.0244,2.3337,2.7116,2.9803
-60,0.67876,1.1620,1.6716,2.0017,2.3011,2.6633,2.9184
-120,0.67660,1.1560,1.6579,1.9802,2.2704,2.6181,2.8608
-inf,0.67449,1.1503,1.6449,1.9600,2.2414,2.5758,2.807
