--- backend/ccfatigue/modules/external_data_sources/astm95.csv
@@ -1,35 +1,0 @@
-n,1,2,3,4,5,6,7,8,9,10,12,15,20,30,40,60,120,inf
-1,161.45,199.5,215.71,224.58,230.16,233.99,236.77,238.88,240.54,241.88,243.91,245.95,248.01,250.09,251.14,252.2,253.25,254.32
-2,18.513,19,19.164,19.247,19.296,19.33,19.353,19.371,19.385,19.396,19.413,19.429,19.446,19.462,19.471,19.479,19.487,19.496
-3,10.128,9.5521,9.2766,9.1172,9.0135,8.9406,8.8868,8.8452,8.8123,8.7855,8.7446,8.7029,8.6602,8.6166,8.5944,8.572,8.5494,8.5265
-4,7.7086,6.9443,6.5914,6.3883,6.256,6.1631,6.0942,6.041,5.9988,5.9644,5.9117,5.8578,5.8025,5.7459,5.717,5.6878,5.6581,5.6281
-5,6.6079,5.7861,5.4095,5.1922,5.0503,4.9503,4.8759,4.8183,4.7725,4.7351,4.6777,4.6188,4.5581,4.4957,4.4658,4.4314,4.3984,4.365
-6,5.9874,5.1433,4.7571,4.5337,4.3874,4.2839,4.2066,4.1468,4.099,4.06,3.9999,3.9381,3.8742,3.8082,3.7743,3.7398,3.7047,3.6688
-7,5.5914,4.7374,4.3468,4.1203,3.9715,3.866,3.787,3.7257,3.6767,3.6365,3.5747,3.5108,3.4445,3.3758,3.3404,3.3043,3.2674,3.2298
-8,5.3177,4.459,4.0662,3.8378,3.6875,3.5806,3.5005,3.4381,3.3881,3.3472,3.284,3.2184,3.1503,3.0794,3.0428,3.0053,2.9669,2.9276
-9,5.1174,4.2565,3.8626,3.6331,3.4817,3.3738,3.2927,3.2296,3.1789,3.1373,3.0729,3.0061,2.9365,2.8637,2.8259,2.7872,2.7475,2.7067
-10,4.9646,4.1028,3.7083,3.478,3.3258,3.2172,3.1355,3.0717,3.0204,2.9782,2.913,2.845,2.774,2.6996,2.6609,2.6211,2.5801,2.5379
-11,4.8443,3.9823,3.5874,3.3567,3.2039,3.0946,3.0123,2.948,2.8962,2.8536,2.7876,2.7186,2.6464,2.5705,2.5309,2.4901,2.448,2.4045
-12,4.7472,3.8853,3.4903,3.2592,3.1059,2.9961,2.9134,2.8486,2.7964,2.7534,2.6866,2.6169,2.5436,2.4663,2.4259,2.3842,2.341,2.2962
-13,4.6672,3.8056,3.4105,3.1791,3.0254,2.9153,2.8321,2.7669,2.7144,2.671,2.6037,2.5331,2.4589,2.3803,2.3392,2.2966,2.2524,2.2064
-14,4.6001,3.7389,3.3439,3.1122,2.9582,2.8477,2.7642,2.6987,2.6458,2.6021,2.5342,2.463,2.3879,2.3082,2.2664,2.223,2.1778,2.1307
-15,4.5431,3.6823,3.2874,3.0556,2.9013,2.7905,2.7066,2.6408,2.5876,2.5437,2.4753,2.4035,2.3275,2.2468,2.2043,2.1601,2.1141,2.0658
-16,4.494,3.6337,3.2389,3.0069,2.8524,2.7413,2.6572,2.5911,2.5377,2.4935,2.4247,2.3522,2.2756,2.1938,2.1507,2.1058,2.0589,2.0096
-17,4.4513,3.5915,3.1968,2.9647,2.81,2.6987,2.6143,2.548,2.4943,2.4499,2.3807,2.3077,2.2304,2.1477,2.104,2.0584,2.0107,1.9604
-18,4.4139,3.5546,3.1599,2.9277,2.7729,2.6613,2.5767,2.5102,2.4563,2.4117,2.3421,2.2686,2.1906,2.1071,2.0629,2.0166,1.9681,1.9168
-19,4.3808,3.5219,3.1274,2.8951,2.7401,2.6283,2.5435,2.4768,2.4227,2.3779,2.308,2.2341,2.1555,2.0712,2.0264,1.9796,1.9302,1.878
-20,4.3513,3.4928,3.0984,2.8661,2.7109,2.599,2.514,2.4471,2.3928,2.3479,2.2776,2.2033,2.1242,2.0391,1.9938,1.9464,1.8963,1.8432
-21,4.3248,3.4668,3.0725,2.8401,2.6848,2.5727,2.4876,2.4205,2.3661,2.321,2.2504,2.1757,2.096,2.0102,1.9645,1.9165,1.8657,1.8117
-22,4.3009,3.4434,3.0491,2.8167,2.6613,2.5491,2.4638,2.3965,2.3419,2.2967,2.2258,2.1508,2.0707,1.9842,1.938,1.8895,1.838,1.7831
-23,4.2793,3.4221,3.028,2.7955,2.64,2.5277,2.4422,2.3748,2.3201,2.2747,2.2036,2.1282,2.0476,1.9605,1.9139,1.8649,1.8128,1.757
-24,4.2597,3.4028,3.0088,2.7763,2.6207,2.5082,2.4226,2.3551,2.3002,2.2547,2.1834,2.1077,2.0267,1.939,1.892,1.8424,1.7897,1.7331
-25,4.2417,3.3852,2.9912,2.7587,2.603,2.4904,2.4047,2.3371,2.2821,2.2365,2.1649,2.0889,2.0075,1.9192,1.8718,1.8217,1.7684,1.711
-26,4.2252,3.369,2.9751,2.7426,2.5868,2.4741,2.3883,2.3205,2.2655,2.2197,2.1479,2.0716,1.9898,1.901,1.8533,1.8027,1.7488,1.6906
-27,4.21,3.3541,2.9604,2.7278,2.5719,2.4591,2.3732,2.3053,2.2501,2.2043,2.1323,2.0558,1.9736,1.8842,1.8361,1.7851,1.7307,1.6717
-28,4.196,3.3404,2.9467,2.7141,2.5581,2.4453,2.3593,2.2913,2.236,2.19,2.1179,2.0411,1.9586,1.8687,1.8203,1.7689,1.7138,1.6541
-29,4.183,3.3277,2.934,2.7014,2.5454,2.4324,2.3463,2.2782,2.2229,2.1768,2.1045,2.0275,1.9446,1.8543,1.8055,1.7537,1.6981,1.6377
-30,4.1709,3.3158,2.9223,2.6896,2.5336,2.4205,2.3343,2.2662,2.2107,2.1646,2.0921,2.0148,1.9317,1.8409,1.7918,1.7396,1.6835,1.6223
-40,4.0848,3.2317,2.8387,2.606,2.4495,2.3359,2.249,2.1802,2.124,2.0772,2.0035,1.9245,1.8389,1.7444,1.6928,1.6373,1.5766,1.5089
-60,4.0012,3.1504,2.7581,2.5252,2.3683,2.254,2.1665,2.097,2.0401,1.9926,1.9174,1.8364,1.748,1.6491,1.5943,1.5343,1.4673,1.3893
-120,3.9201,3.0718,2.6802,2.4472,2.29,2.175,2.0867,2.0164,1.9588,1.9105,1.8337,1.7505,1.6587,1.5543,1.4952,1.429,1.3519,1.2539
-inf,3.8415,2.9957,2.6049,2.3719,2.2141,2.0986,2.0096,1.9384,1.8799,1.8307,1.7522,1.6664,1.5705,1.4591,1.394,1.318,1.2214,1
