--- backend/ccfatigue/modules/external_data_sources/astm99.csv
@@ -1,35 +1,0 @@
-﻿n,1,2,3,4,5,6,7,8,9,10,12,15,20,30,40,60,120,inf
-1,4052.2,4999.5,5403.3,5624.6,5763.7,5859,5928.3,5981.6,6022.5,6055.8,6106.3,6157.3,6208.7,6260.7,6286.8,6313,6339.4,6366
-2,98.503,99,99.166,99.249,99.299,99.332,99.356,99.374,99.388,99.399,99.416,99.432,99.449,99.466,99.474,99.483,99.491,99.501
-3,34.116,30.817,29.457,28.71,28.237,27.911,27.672,27.489,27.345,27.229,27.052,26.872,26.69,26.505,26.411,26.316,26.221,26.125
-4,21.198,18,16.694,15.977,15.522,15.207,14.976,14.799,14.659,14.546,14.374,14.198,14.02,13.838,13.745,13.652,13.558,13.463
-5,16.258,13.274,12.06,11.392,10.967,10.672,10.456,10.289,10.158,10.051,9.8883,9.7222,9.5527,9.3793,9.2912,9.202,9.1118,9.0204
-6,13.745,10.925,9.7795,9.1483,8.7459,8.4661,8.26,8.1016,7.9761,7.8741,7.7183,7.559,7.3958,7.2285,7.1432,7.0568,6.969,6.8801
-7,12.246,9.5466,8.4513,7.8467,7.4604,7.1914,6.9928,6.8401,6.7188,6.6201,6.4691,6.3143,6.1554,5.9921,5.9084,5.8236,5.7372,5.6495
-8,11.259,8.6491,7.591,7.006,6.6318,6.3707,6.1776,6.0289,5.9106,5.8143,5.6668,5.5151,5.3591,5.1981,5.1156,5.0316,4.946,4.8588
-9,10.561,8.0215,6.9919,6.4221,6.0569,5.8018,5.6129,5.4671,5.3511,5.2565,5.1114,4.9621,4.808,4.6486,4.5667,4.4831,4.3978,4.3105
-10,10.044,7.5594,6.5523,5.9943,5.6363,5.3858,5.2001,5.0567,4.9424,4.8492,4.7059,4.5582,4.4054,4.2469,4.1653,4.0819,3.9965,3.909
-11,9.646,7.2057,6.2167,5.6683,5.316,5.0692,4.8861,4.7445,4.6315,4.5393,4.3974,4.2509,4.099,3.9411,3.8596,3.7761,3.6904,3.6025
-12,9.3302,6.9266,5.9526,5.4119,5.0643,4.8206,4.6395,4.4994,4.3875,4.2961,4.1553,4.0096,3.8584,3.7008,3.6192,3.5355,3.4494,3.3608
-13,9.0738,6.701,5.7394,5.2053,4.8616,4.6204,4.441,4.3021,4.1911,4.1003,3.9603,3.8154,3.6646,3.507,3.4253,3.3413,3.2548,3.1654
-14,8.8616,6.5149,5.5639,5.0354,4.695,4.4558,4.2779,4.1399,4.0297,3.9394,3.8001,3.6557,3.5052,3.3476,3.2656,3.1813,3.0942,3.004
-15,8.6831,6.3589,5.417,4.8932,4.5556,4.3183,4.1415,4.0045,3.8948,3.8049,3.6662,3.5222,3.3719,3.2141,3.1319,3.0471,2.9595,2.8684
-16,8.531,6.2262,5.2922,4.7726,4.4374,4.2016,4.0259,3.8896,3.7804,3.6909,3.5527,3.4089,3.2588,3.1007,3.0182,2.933,2.8447,2.7528
-17,8.3997,6.1121,5.185,4.669,4.3359,4.1015,3.9267,3.791,3.6822,3.5931,3.4552,3.3117,3.1615,3.0032,2.9205,2.8348,2.7459,2.653
-18,8.2854,6.0129,5.0919,4.579,4.2479,4.0146,3.8406,3.7054,3.5971,3.5082,3.3706,3.2273,3.0771,2.9185,2.8354,2.7493,2.6597,2.566
-19,8.185,5.9259,5.0103,4.5003,4.1708,3.9386,3.7653,3.6305,3.5225,3.4338,3.2965,3.1533,3.0031,2.8442,2.7608,2.6742,2.5839,2.4893
-20,8.096,5.8489,4.9382,4.4307,4.1027,3.8714,3.6987,3.5644,3.4567,3.3682,3.2311,3.088,2.9377,2.7785,2.6947,2.6077,2.5168,2.4212
-21,8.0166,5.7804,4.874,4.3688,4.0421,3.8117,3.6396,3.5056,3.3981,3.3098,3.1729,3.0299,2.8796,2.72,2.6359,2.5484,2.4568,2.3603
-22,7.9454,5.719,4.8166,4.3134,3.988,3.7583,3.5867,3.453,3.3458,3.2576,3.1209,2.978,2.8274,2.6675,2.5831,2.4951,2.4029,2.3055
-23,7.8811,5.6637,4.7649,4.2635,3.9392,3.7102,3.539,3.4057,3.2986,3.2106,3.074,2.9311,2.7805,2.6202,2.5355,2.4471,2.3542,2.2559
-24,7.8229,5.6136,4.7181,4.2184,3.8951,3.6667,3.4959,3.3629,3.256,3.1681,3.0316,2.8887,2.738,2.5773,2.4923,2.4035,2.3099,2.2107
-25,7.7698,5.568,4.6755,4.1774,3.855,3.6272,3.4568,3.3239,3.2172,3.1294,2.9931,2.8502,2.6993,2.5383,2.453,2.3637,2.2695,2.1694
-26,7.7213,5.5263,4.6366,4.14,3.8183,3.5911,3.421,3.2884,3.1818,3.0941,2.9579,2.815,2.664,2.5026,2.417,2.3273,2.2325,2.1315
-27,7.6767,5.4881,4.6009,4.1056,3.7848,3.558,3.3882,3.2558,3.1494,3.0618,2.9256,2.7827,2.6316,2.4699,2.384,2.2938,2.1984,2.0965
-28,7.6356,5.4529,4.5681,4.074,3.7539,3.5276,3.3581,3.2259,3.1195,3.032,2.8959,2.753,2.6017,2.4397,2.3535,2.2629,2.167,2.0642
-29,7.5976,5.4205,4.5378,4.0449,3.7254,3.4995,3.3302,3.1982,3.092,3.0045,2.8685,2.7256,2.5742,2.4118,2.3253,2.2344,2.1378,2.0342
-30,7.5625,5.3904,4.5097,4.0179,3.699,3.4735,3.3045,3.1726,3.0665,2.9791,2.8431,2.7002,2.5487,2.386,2.2992,2.2079,2.1107,2.0062
-40,7.3141,5.1785,4.3126,3.8283,3.5138,3.291,3.1238,2.993,2.8876,2.8005,2.6648,2.5216,2.3689,2.2034,2.1142,2.0194,1.9172,1.8047
-60,7.0771,4.9774,4.1259,3.6491,3.3389,3.1187,2.953,2.8233,2.7185,2.6318,2.4961,2.3523,2.1978,2.0285,1.936,1.8363,1.7263,1.6006
-120,6.851,4.7865,3.9493,3.4796,3.1735,2.9559,2.7918,2.6629,2.5586,2.4721,2.3363,2.1915,2.0346,1.86,1.7628,1.6557,1.533,1.3805
-inf,6.6349,4.6052,3.7816,3.3192,3.0173,2.802,2.6393,2.5113,2.4073,2.3209,2.1848,2.0385,1.8783,1.6964,1.5923,1.473,1.3246,1
