--- backend/fortran/__init__.py
@@ -1,4 +1,0 @@
-"""
-Folder for fortran codes
-Will be compiled to python with f2py to ccfatigue.fortran
-"""
